+++ Ryujinx.Ava/UI/Windows/TitleUpdateWindow.axaml
@@ -1,22 +1,23 @@
-<window:StyleableWindow
+<UserControl
     x:Class="Ryujinx.Ava.UI.Windows.TitleUpdateWindow"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:locale="clr-namespace:Ryujinx.Ava.Common.Locale"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:window="clr-namespace:Ryujinx.Ava.UI.Windows"
+    xmlns:viewModels="clr-namespace:Ryujinx.Ava.UI.ViewModels"
+    xmlns:models="clr-namespace:Ryujinx.Ava.UI.Models"
     Width="600"
     Height="400"
     MinWidth="600"
     MinHeight="400"
     MaxWidth="600"
     MaxHeight="400"
-    SizeToContent="Height"
-    WindowStartupLocation="CenterOwner"
     mc:Ignorable="d"
+    x:CompileBindings="True"
+    x:DataType="viewModels:TitleUpdateViewModel"
     Focusable="True">
-    <Grid Margin="15">
+    <Grid Margin="10">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
@@ -24,22 +25,20 @@
             <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
         <TextBlock
-            Name="Heading"
             Grid.Row="1"
-            MaxWidth="500"
-            Margin="20,15,20,20"
             HorizontalAlignment="Center"
             VerticalAlignment="Center"
-            LineHeight="18"
-            TextAlignment="Center"
-            TextWrapping="Wrap" />
+            TextAlignment="Left"
+            TextWrapping="Wrap"
+            Text="{Binding Heading}" />
         <Border
             Grid.Row="2"
-            Margin="5"
+            Margin="0 10"
             HorizontalAlignment="Stretch"
             VerticalAlignment="Stretch"
             BorderBrush="Gray"
-            BorderThickness="1">
+            BorderThickness="1"
+            CornerRadius="5">
             <ScrollViewer
                 VerticalAlignment="Stretch"
                 HorizontalScrollBarVisibility="Auto"
@@ -48,9 +47,10 @@
                     Margin="10"
                     HorizontalAlignment="Stretch"
                     VerticalAlignment="Stretch"
-                    Items="{Binding _titleUpdates}">
+                    Items="{Binding TitleUpdates}">
                     <ItemsControl.ItemTemplate>
-                        <DataTemplate>
+                        <DataTemplate
+                            x:DataType="models:TitleUpdateModel">
                             <RadioButton
                                 Padding="8,0"
                                 VerticalContentAlignment="Center"
@@ -106,10 +106,10 @@
                     Name="CancelButton"
                     MinWidth="90"
                     Margin="5"
-                    Command="{Binding Close}">
+                    Click="Close">
                     <TextBlock Text="{locale:Locale InputDialogCancel}" />
                 </Button>
             </DockPanel>
         </DockPanel>
     </Grid>
-</window:StyleableWindow>
+</UserControl>
